Restore bats-support as explicit devDep
It was removed because npm 7+ now installs peer deps by default, and
bats-support is a peerDep of bats-assert.

However, that implies a version restriction that this plugin depends on
npm 7+. But that restriction only applies to _development_. We could put
the restriction on bats-assert, (which would be more aligned with it as
a dev dep). But it's not technically true because any dependent of
bats-assert could avoid the npm restriction by simply installing
bats-support _explicitly_.

Which leads us back to the main point. bats-support is loaded _from_ the
parent dependent. So it should be an explicit devDep at that location.